marl_research/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "marl_research"
version = "0.1.0"
description = "Multi-Agent Reinforcement Learning Research Template"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Anonymous", email = "anonymous@example.com"}
]
dependencies = [
    "numpy>=1.21.0",
    "torch>=1.12.0",
    "tensorboard>=2.10.0",
    "gymnasium>=0.28.0",
    "pyyaml>=6.0",
    "wandb>=0.13.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "tqdm>=4.64.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.12.0",
]

[project.optional-dependencies]
smac = ["pysc2>=3.0.0", "smac>=1.0.0"]
smacv2 = ["smacv2>=1.0.0"]
overcooked = ["overcooked-ai>=1.1.0"]
dev = ["pytest>=7.0.0", "black>=22.0.0", "isort>=5.10.0", "flake8>=5.0.0"]
all = ["marl_research[smac,smacv2,overcooked,dev]"]

[project.scripts]
marl-train = "marl_research.runners.train:main"
marl-eval = "marl_research.runners.evaluate:main"
vabl-experiments = "marl_research.scripts.run_vabl_experiments:main"
vabl-quick-test = "marl_research.scripts.run_quick_test:main"
vabl-ablation = "marl_research.scripts.run_ablation_study:main"
vabl-long-run = "marl_research.scripts.run_long_comparison:main"

[tool.setuptools.packages.find]
where = ["."]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310']

[tool.isort]
profile = "black"
line_length = 100
